feat: 랜딩페이지 탭상수 분리 (#29)
