+++ src/pages/HomePage.tsx
@@ -1,42 +1,11 @@
 import { useState } from 'react'
-import MainExam from '@/assets/images/main-exam.png'
-import MainQna from '@/assets/images/main-qna.png'
-import MainCommunity from '@/assets/images/main-community.png'
 import MainBanner from '@/assets/images/main-banner.png'
-
-type TabType = '쪽지시험' | '질의응답' | '커뮤니티'
-
-const TAB_LIST: TabType[] = ['쪽지시험', '질의응답', '커뮤니티']
-
-type TabContent = {
-  title: [string, string]
-  image: string
-  alt: string
-}
-
-type TabContents = {
-  쪽지시험: TabContent
-  질의응답: TabContent
-  커뮤니티: TabContent
-}
-
-const TAB_CONTENTS: TabContents = {
-  쪽지시험: {
-    title: ['쪽지시험으로', '실력을 차곡차곡 쌓아보세요'],
-    image: MainExam,
-    alt: '쪽지시험 서비스 화면 예시',
-  },
-  질의응답: {
-    title: ['질문하고 배우고', '동료 수강생과 함께 성장해요'],
-    image: MainQna,
-    alt: '질의응답 서비스 화면 예시',
-  },
-  커뮤니티: {
-    title: ['정보 공유부터 팀원 모집까지', '커뮤니티에서 함께 해결해요'],
-    image: MainCommunity,
-    alt: '커뮤니티 서비스 화면 예시',
-  },
-}
+import {
+  type TabType,
+  TAB_LIST,
+  TAB_CONTENTS,
+} from '@/constants/mainTabContents'
+import { cn } from '@/lib/utils'
 
 function HomePage() {
   const [tab, setTab] = useState<TabType>('쪽지시험')
@@ -67,11 +36,13 @@ function HomePage() {
                 <button
                   type="button"
                   onClick={() => setTab(item)}
-                  className={`flex h-[54px] w-[143px] items-center justify-center rounded-[40px] ${
-                    tab === item
-                      ? 'bg-btn-fill-default text-grey-1'
-                      : 'text-grey-9'
-                  }`}
+                  className={cn(
+                    `flex h-[54px] w-[143px] items-center justify-center rounded-[40px] ${
+                      tab === item
+                        ? 'bg-btn-fill-default text-grey-1'
+                        : 'text-grey-9'
+                    }`
+                  )}
                 >
                   {item}
                 </button>
